* fix bug in TarArchive implementation. Bug details given below.
When the file type indicated by the archive (which, in the case of UNIX V6 tar, can be
hard link, symlink, and normal file) is not "normal file," and the length of this entry as
given in the header is zero, the parser would skip ahead 512 bytes (the tar block size).

As the entry was zero bytes, it consisted of only a header. This means that skipping
ahead 512 bytes skips past the header for the next entry and in to the entry's contents
(if present). This means that the parser attempts to parse the entry's contents as a tar
header, and will fail (throwing an InvalidDataException) or misbehave if the entry's data
happens, by chance, to have a valid sequence of octal digits in the right location.