fix bug in Routing#GetParameters (see below).
Unbounded parameters have an asterisk before their names. GetParameters did not take this in to account and, as a result, using it on a format string with unbounded parameters resulted in the asterisk being counted as part of the variable's name. This causes methods which rely on a specific parameter being present (such as Archiving#AddArchive) to throw an exception if said parameter is unbounded.