adds support for HTTP request pipelining (closes #29)
Turns out it was a fairly simple addition. HTTP pipelining uses the same TCP connection to send multiple requests, so all we need to do is check whether there's still data after processing one request. If there's still data, we just need to run it through the function again (making sure to close the HttpResponse object so we dispose of any buffers/send any buffered data). The only problem with this is if so many requests are pipelined that it causes a stack overflow. This is an unlikely situation.