adjust HttpResponse transmit buffer size
Testing revealed that a 16kB buffer provided significant improvements over a 4kB buffer. The test used a 2.3MB JPEG and the text string "Hello, World!" and compared the response times, using the fastest response time over 4 requests. If the results were not consistent, the test was redone. The results are below.

|  Buffer  |  Text (ms)  |  Image (ms)  |
|----------|-------------|--------------|
|  4096    |  0.62       |   97.5       |
|  8192    |  0.64       |   90.7       |
|  16384   |  0.65       |   78.6       |
|----------|-------------|--------------|