add SingleFileArchive abstract class
The main Archive class assumes that there will be a plurality of files. While this is often the case with archives, Zener already contains two Archive subclasses that will only ever contain a single file. To reduce code duplication, and provide properties that are more useful when an archive will only ever contain a single file, the SingleFileArchive class should be used.