revise HttpRequest#DeterminePostHandler
Switches around order of comparison. This makes it work with
multipart/form-data. The reason for this is that if you call
IsEquivalent on a media type with parameters, IsEquivalent will require
the media type passed to IsEquivalent to have those parameters. With
multipart/form-data, where the client sends a boundary parameter, this
doesn't work unless you switch them around.